Move testsuites css out of library and into testsuites
