+++ src/ui/components/Library/Team/View/TestRuns/TestRunListItem.tsx
@@ -9,7 +9,7 @@ import { getTruncatedRelativeDate } from "../Recordings/RecordingListItem/Record
 import { AttributeContainer } from "./AttributeContainer";
 import { ModeAttribute } from "./Overview/RunSummary";
 import { TestRunsContext } from "./TestRunsContextRoot";
-import styles from "../../../Library.module.css";
+import styles from "../../../Testsuites.module.css";
 
 function Status({ failCount }: { failCount: number }) {
   const status = failCount > 0 ? "fail" : "success";
